generate-repology-metadata: add helper functions for json printing
To make it easier to change and extend the information we collect for
repology.  Updated script almost behaves like previous version, except
that the depends array is now printed with one entry per line.

srcurl is still omitted if it is empty, though maybe it would be
better to include it and set it to "" if that is the case.