generate-repology-metadata: include build recipe and patch urls
Both raw and "normal" links are included. Urls are generated based on
git upstream remote, and assumed github style when it comes to the raw
url format, so if origin remote doesn't exist, or tracks something
other than github, then the script will fail.

Some .debug patches, and some .diff files that are "manually" applied,
are not included in patch list.