Include commit hashes in URLs instead of branch name (master)
Currently Repology Metadata is generated every 6 hours, so during the
moment a package is updated in repo and new Metadata is generated, there
is maximum delay of 6 hours. During these 6 hours, the updated patch may
fail to apply for other people trying to apply patch from master which
was indeed for other version. Also this will make the metadata quite
history proof.