Revert "chore: clone with 1 depth"
We need a full clone to ensure that we link to the patches with their
last commit in the metadata. The last few commits are having a large
diff because --depth=1, the metadata just contained the latest commit as
the commit for the generated URLs for the patches. This behaviour is not
desired

This reverts commit 27bab4e2f128173f15038df64f892854b25a6314.